chore: Move the auth header generation classes to the types.py file and make them a Protocol instead of a Union
the goal of this is to remove things from generic 'utils' style junk
drawers and have them be clearly labeled
